requirements.txt
gitpython
streamlit
tree_sitter
jaseci
